Move some legacy comps to `uis`
