+++ src/components/pages/Comunity.svelte
@@ -1,0 +1,12 @@
+<style lang="postcss">
+  .container {
+    display: grid;
+
+    background-color: var(--background);
+  }
+</style>
+
+<script lang="ts">
+</script>
+
+<div class="container">Text</div>
